+++ my-app/src/frontend/components/index.js
@@ -1,0 +1,4 @@
+export { Filter } from "./filter/Filter";
+export { Navbar } from "./navbar/Navbar";
+export { Products } from "./products/Products";
+export { Footer } from "./footer/Footer";
