+++ my-app/src/frontend/pages/signup/Signup.js
@@ -3,7 +3,7 @@ import "../../../App.css";
 import { useAuth } from "../../contexts/index";
 import axios from "axios";
 import { Link, useNavigate } from "react-router-dom";
-import { Footer } from "../../components/footer/Footer";
+import { Footer } from "../../components/index";
 
 export function Signup() {
   const { authState, authDispatch } = useAuth();
